requirements.txt
APScheduler==3.11.0
Authlib==1.4.0
blinker==1.9.0
Brotli==1.1.0
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
config==0.5.1
cryptography==44.0.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
plotly==5.24.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
zstandard==0.23.0
